refactor(core): extract loop input, backend and display boundaries
